Initial Commit
This represents the initial commit for the source of the engine.

As of now, currently, there is a basic window factory implemented, and the required functionality for creating a window using the Win32 API. Some of the codes are reused from my previous engine.

After getting more things working, I will switch to CMake for the build system in order to move on other platforms easier.