Multiple functionality update
Removed the VS project, because the engine got reworked to use CMake out of the box.

glad2, glm and imgui dependencies were added to the engine internally.
Added functionality for Windows, Linux and Android(still WIP, but JNI works this time!)
Abstracted many things related to graphical backends, contexts and window management.

For now, only rendering works. Input management will be taken care of in future updates of the engine.